Move Panel View tests
